Do not create gfortran symlink on macOS
The original CI workflow assumed that an executable `gfortran-14` was available on macOS, but not a `gfortran`. Therefore a symlink named `gfortran` was created without even checking if this was really necessary. Recent CI failures showed instead that such a file already existed.

Therefore, we now use the available command `gfortran` on macOS as well as on Ubuntu, with no distinction.