Do not create gfortran symlink on macOS (#23)
The original CI workflow assumed that an executable `gfortran-14` was
available on macOS, but not a `gfortran`. Therefore a symlink named
`gfortran` was created without even checking if this was really
necessary. Recent CI failures showed instead that a binary file named
`gfortran` already existed.

Therefore, we now use the available command `gfortran` on macOS as well
as Ubuntu, with no distinction. Fixes #22.